Added [feature/navigation-refactor]: Separated navigation logic into dedicated files
